+++ nym-vpn-core/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - While in Connected state swap internal resolver to use custom DNS (via system resolver). (https://github.com/nymtech/nym-vpn-client/pull/5674)
+- Use "Auto" for entry and exit selectors independently (https://github.com/nymtech/nym-vpn-client/pull/5962)
 
 ### Fixed